+++ src/environments/environments/CarTrackParentEnvironment.py
@@ -14,7 +14,9 @@ from sensor_msgs.msg import LaserScan
 from environment_interfaces.srv import Reset
 from message_filters import Subscriber, ApproximateTimeSynchronizer
 
-class CarTrackParentEnvironment(Node):
+from environments.ParentCarEnvironment import ParentCarEnvironment
+
+class CarTrackParentEnvironment(ParentCarEnvironment):
     """
     CarTrack Reinforcement Learning Environment:
 
@@ -29,8 +31,8 @@ class CarTrackParentEnvironment(Node):
         
         Reward:
             It's progress toward the goal plus,
-            100+ if it reaches the goal plus,
-            -50 if it collides with the wall
+            50+ if it reaches the goal plus,
+            -25 if it collides with the wall
 
         Termination Conditions:
             When the agent is within REWARD_RANGE units or,
@@ -41,61 +43,15 @@ class CarTrackParentEnvironment(Node):
     """
 
     def __init__(self, car_name, reward_range=1, max_steps=50, collision_range=0.2, step_length=0.5):
-        super().__init__('car_track_environment')
-        
+        super().__init__('car_track', car_name, reward_range, max_steps, collision_range, step_length)
+
         # Environment Details ----------------------------------------
-        self.NAME = car_name
-        self.REWARD_RANGE = reward_range
-        self.MAX_STEPS = max_steps
         self.MAX_STEPS_PER_GOAL = max_steps
-        self.COLLISION_RANGE = collision_range
-        self.STEP_LENGTH = step_length
-
-        self.MAX_ACTIONS = np.asarray([3, 3.14])
         self.MIN_ACTIONS = np.asarray([0, -3.14])
-        self.OBSERVATION_SIZE = 8 + 10  # Car position + Lidar rays + goal position
-        self.ACTION_NUM = 2
-
-
-        # Pub/Sub ----------------------------------------------------
-        self.cmd_vel_pub = self.create_publisher(
-                Twist,
-                f'/{self.NAME}/cmd_vel',
-                10
-            )
-        
-        self.odom_sub = Subscriber(
-            self,
-            Odometry,
-            f'/{self.NAME}/odometry',
-        )
-        # TODO: Map the lidar to a dynamic topic => of the form /model/<name>/lidar
-        self.lidar_sub = Subscriber(
-            self,
-            LaserScan,
-            f'/lidar',
-        )
-
-        self.message_filter = ApproximateTimeSynchronizer(
-            [self.odom_sub, self.lidar_sub],
-            10,
-            0.1,
-        )
-
-        self.message_filter.registerCallback(self.message_filter_callback)
-
-        self.observation_future = Future()
+        self.OBSERVATION_SIZE = 8 + 10  # Car position + Lidar rays
 
         # Reset Client -----------------------------------------------
-        self.reset_client = self.create_client(
-            Reset,
-            'car_track_reset'
-        )
-
-        while not self.reset_client.wait_for_service(timeout_sec=1.0):
-            self.get_logger().info('reset service not available, waiting again...')
-
-        self.goal_position = [10, 10]  # x and y
+        self.goal_number = 0
         self.all_goals = []
 
         self.car_reset_positions = {
@@ -104,15 +60,10 @@ class CarTrackParentEnvironment(Node):
             'yaw': 0.0
         }
 
-        self.timer = self.create_timer(step_length, self.timer_cb)
-        self.timer_future = Future()
         self.step_counter = 0
 
         self.get_logger().info('Environment Setup Complete')
     
-    def timer_cb(self):
-        self.timer_future.set_result(True)
-
     def reset(self):
         self.step_counter = 0
 
@@ -139,28 +90,6 @@ class CarTrackParentEnvironment(Node):
         print("Goal", number, "spawned")
         return self.all_goals[number % len(self.all_goals)]
 
-    def step(self, action):
-        self.step_counter += 1
-
-        state = self.get_observation()
-
-        lin_vel, ang_vel = action
-        self.set_velocity(lin_vel, ang_vel)
-
-        while not self.timer_future.done():
-            rclpy.spin_once(self)
-
-        self.timer_future = Future()
-        
-        next_state = self.get_observation()
-        reward = self.compute_reward(state, next_state)
-        terminated = self.is_terminated(next_state)
-        print(self.step_counter, self.MAX_STEPS)
-        truncated = self.step_counter >= self.MAX_STEPS
-        info = {}
-
-        return next_state, reward, terminated, truncated, info
-
     def call_reset_service(self):
         x, y = self.goal_position
 
@@ -171,10 +100,10 @@ class CarTrackParentEnvironment(Node):
         request.cy = self.car_reset_positions['y']
         request.cyaw = self.car_reset_positions['yaw']
         request.flag = "car_and_goal"
+
         future = self.reset_client.call_async(request)
         rclpy.spin_until_future_complete(self, future)
 
-        # print(f'Reset Response Recieved: {future.result()}')
         return future.result()
 
     def update_goal_service(self, number):
@@ -200,7 +129,7 @@ class CarTrackParentEnvironment(Node):
         # ranges, _ = self.process_lidar(lidar)
 
         reduced_range = self.sample_reduce_lidar(lidar)
-        # print(reduced_range)
+
         # Get Goal Position
         return odom + reduced_range 
 
@@ -211,7 +140,7 @@ class CarTrackParentEnvironment(Node):
             -1 to -2 => goal x, y
             9 to -3 => lidar
         """
-        collided_wall = self.has_collided(observation[9:-2])
+        collided_wall = self.has_collided(observation)
         flipped_over = self.has_flipped_over(observation)
 
         if collided_wall:
@@ -220,13 +149,6 @@ class CarTrackParentEnvironment(Node):
             print("Flipped over")
 
         return collided_wall or flipped_over
-    
-    def has_flipped_over(self, observation):
-        w, x, y, z = observation[2:6]
-        return abs(x) > 0.5 or abs(y) > 0.5
-    
-    def has_collided(self, lidar_ranges):
-        return any(0 < ray < self.COLLISION_RANGE for ray in lidar_ranges)
 
     def compute_reward(self, state, next_state):
 
@@ -252,34 +174,10 @@ class CarTrackParentEnvironment(Node):
             self.step_counter = 0
             self.update_goal_service(self.goal_number)
             
-        if self.has_collided(next_state[9:-2]) or self.has_flipped_over(next_state):
+        if self.has_collided(next_state) or self.has_flipped_over(next_state):
             reward -= 25 # TODO: find optimal value for this
         
-        
-        # reward += delta_distance
-
         return reward
-
-    def message_filter_callback(self, odom: Odometry, lidar: LaserScan):
-        self.observation_future.set_result({'odom': odom, 'lidar': lidar})
-
-    def get_data(self):
-        rclpy.spin_until_future_complete(self, self.observation_future)
-        future = self.observation_future
-        self.observation_future = Future()
-        data = future.result()
-        return data['odom'], data['lidar']
-    
-    def process_odom(self, odom: Odometry):
-        pose = odom.pose.pose
-        position = pose.position
-        orientation = pose.orientation
-
-        twist = odom.twist.twist
-        lin_vel = twist.linear
-        ang_vel = twist.angular
-
-        return [position.x, position.y, orientation.w, orientation.x, orientation.y, orientation.z, lin_vel.x, ang_vel.z]
 
     def process_lidar(self, lidar: LaserScan):
         ranges = lidar.ranges
@@ -304,15 +202,3 @@ class CarTrackParentEnvironment(Node):
             reduced_range.append(sample)
 
         return reduced_range
-
-    def set_velocity(self, linear, angular):
-        """
-        Publish Twist messages to f1tenth cmd_vel topic
-        """
-        velocity_msg = Twist()
-        velocity_msg.angular.z = float(angular)
-        velocity_msg.linear.x = float(linear)
-
-        self.cmd_vel_pub.publish(velocity_msg)
-
-
